[refactor] move all forms queries to separate file and rewrite for clarity.
