+++ src/supabase/queries/cases.ts
@@ -162,120 +162,6 @@ export function formatPartialCaseFromQuery(item: any): CasePartial {
 }
 
 /**
- * Fetch the supabase storage publicUrl associated with the given form filename.
- *
- * @param filename name of form file stored in supabase storage
- * @returns publicUrl of the form
- */
-export async function getPublicFormUrl(filename: string): Promise<string> {
-  try {
-    // fetch the form objects from supabase storage
-    const { data } = await supabase.storage
-      .from('caseFiles')
-      .getPublicUrl(filename);
-    return data.publicUrl;
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.warn('(getFormsByCaseUid)', error);
-    throw error;
-  }
-}
-
-export async function fetchFormByName(
-  caseUid: CaseUid,
-  filename: string,
-): Promise<Form> {
-  try {
-    // fetch rows with the matching CaseUid
-    const { data } = await supabase
-      .from('caseForms')
-      .select()
-      .eq('caseId', caseUid)
-      .eq('filename', filename);
-
-    if (!data) {
-      throw new Error(`no forms found for the given case: ${caseUid}`);
-    }
-
-    return await formatForm(caseUid, data[0]);
-    // return data.map(item => {
-    //   const form: FormPartial = {
-    //     id: item.formId,
-    //     title: item.title,
-    //     filename: item.filename,
-    //     date: item.date,
-    //   };
-    //   return form;
-    // });
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.warn('(getFormObjects)', error);
-    throw error;
-  }
-}
-
-export async function formatForm(caseUid: CaseUid, item: any): Promise<Form> {
-  const partialForm = formatPartialFormFromQuery(item);
-  const formUrl = await getPublicFormUrl(`${caseUid}/${partialForm.filename}`);
-  const formData: Form = {
-    ...partialForm,
-    formUrl,
-  };
-  return formData;
-}
-/**
- * Parse supabase case query and return `CasePartial` object.
- *
- * @param item supabase Case query return data
- * @returns `CasePartial` object
- */
-export function formatPartialFormFromQuery(item: any): FormPartial {
-  const formattedPartial: FormPartial = {
-    id: item.caseId,
-    title: item.title,
-    filename: item.string,
-    date: item.date,
-  };
-  return formattedPartial;
-}
-
-/**
- * Fetch the `FormPartial` for forms associated with a given case.
- *
- * @param caseUid uid of target case.
- * @returns list of `FormPartial` objects.
- */
-export async function getPartialForms(
-  caseUid: CaseUid,
-): Promise<FormPartial[]> {
-  try {
-    // fetch rows with the matching CaseUid
-    const { data } = await supabase
-      .from('caseForms')
-      .select()
-      .eq('caseId', caseUid);
-
-    if (!data) {
-      throw new Error(`no forms found for the given case: ${caseUid}`);
-    }
-
-    return data.map(item => {
-      const form: FormPartial = {
-        id: item.formId,
-        title: item.title,
-        filename: item.filename,
-        date: item.date,
-      };
-      return form;
-    });
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.warn('(getFormObjects)', error);
-    throw error;
-  }
-}
-
-/**
  * Update a specific User/Case status
  *
  * @param caseId specified caseId
